US1460: Moved code for preventing URLEncoding from routes.js in app to crds_utilities.js in core.
