--- crossroads.net/core/utilities/cookies.js
@@ -1,15 +1,0 @@
-'use strict()';
-
-module.exports = {
-  getCookie: function(cname) {
-    var name = cname + "=";
-    var ca = document.cookie.split(';');
-    for (var i = 0; i < ca.length; i++) {
-      var c = ca[i];
-      while (c.charAt(0) == ' ') c = c.substring(1);
-      if (c.indexOf(name) == 0) return c.substring(name.length, c.length);
-    }
-    return "";
-  },
-}
-
